Improve TriangularView component
Add extension to read attributes for the view
Add Overload constructor to the TriangularView
Remove path variable
Refactor path1 to path
