+++ app/src/main/res/layout/card_item_pokemon_type.xml
@@ -33,7 +33,7 @@
             app:layout_constraintTop_toTopOf="parent"
             tools:text="TYPE"/>
 
-        <mx.yellowme.youst.common.components.TriangularView
+        <mx.yellowme.youst.components.TriangularView
                 android:id="@+id/typeDecoration"
                 android:layout_width="0dp"
                 android:layout_weight="0.5"
